# [FIX] point_of_sale: reduce timeout/memory errors likelihood when loading up a pos session