Loading `product.pricelist.items` when opening a pos session can lead to both Timeout and Memory Errors
if the number of items in a customer DB is too big. This PR adds two commits to tackle this issue.

##### First commit

The first commit adds two optimizations. The first one is to set the `fields` property on the pos `product.pricelist.item`
definition. Not only does this avoid recomputing unused fields (rule_tip, price, name), it also prevents the recomputation
of non-stored computed fields that the customer might have added through studio.

The second optimization is to add a `readLoad` property to pos models. This property turns off the `name_get` in 
`_read_format`. In other words, instead of returning (id, display_name) for many2one, `_read` only returns id. 
This is helpful for many2one to models whose name_get/display_name computations is expensive (e.g. `product.product`).

Both optimizations have the side-effect of reducing the `search_read` memory footprint.

##### Second commit

The second commit adds a `_paginatedRPC` feature to point_of_sale models to paginate
rpc calls to search_read/read. It also adds a new property on models called `pageSize` that turns on this new feature.
The idea is to avoid MemoryErrors by doing multiple rpc calls returning fixed-size result sets.

The limited/load in background design used for `res.partner` and `product.product` cannot be used here because
`product.product` depends on `product.pricelist.item`. I.e. the corresponding pricelist.item must be loaded before
a `product.product` is shown in the pos. Since we cannot guarantee that loading the first 20 000 `product.pricelist.items`
will match the first 20 000 `product.products`, we need to load all the items beforehand.

Ideally the `pageSize` property should be set on the pos.config but adding a stored field is not allowed in stable.

Now I'm not really sure that this feature is safe enough to be deployed in stable, hence being in a separate commit
should it be dropped/cherry-pick from this branch.


#### Speedup

Customer database with 463 272 `product.pricelist.item`.
Timing to load all the pricelist items for the given session, increasing the number of items.
Paginated rpc pageSize = 30 000, timing = sum of paginated rpc timings.

| Number of items | Before PR | After PR | After PR (paginated rpc) |
|:-----------------:|:-----------:|:--------:|:-------------------------:|
| 500 | 410ms | 119ms | 119ms|
| 2 000 | 953ms | 185ms | 185ms |
| 10 000 | 4.36s | 602ms | 602ms |
| 30 000 | 13.03s | 1.43s | 1.6s |
| 100 000 | 47.38s | 4.88s | 5.3s |
| 463 272 | 3.5min<sup>1</sup> | 21.42s | 37.13s |

1: MemoryError raised with default odoo config

Search_read results size, paginated rpc pageSize = 30 000, size = max size of a single call's results.
| Number of items | Before PR | After PR | After PR (paginated rpc) |
|:------------------:|:----------:|:---------:|:------------------------:|
| 500 | 553 Kb  | 203 Kb | 203 Kb|
| 2000 | 2.2 Mb | 811 kb | 811 Kb |
| 10 000 | 10.9 Mb | 4.1 Mb | 4.1 Mb |
| 30 000 | 33.1 Mb | 12.2 Mb | 12.2 Mb |
| 100 000 | 110 Mb | 40 Mb | 12.2 Mb |
| 463 272 | 510 Mb<sup>1</sup>| 188 Mb | 12.2 Mb|

1: MemoryError raised with default odoo config.

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr