@Aurelienvd 2023-05-22
@odoo/pos Hello, could someone have a look at this PR when they have the time ? Many thanks !